requirements.txt
dspy
pydantic
pydantic-ai
python-dotenv
